+++ projects/igniteui-angular-i18n/src/i18n/BG/combo-resources.ts
@@ -3,7 +3,7 @@ import { IComboResourceStrings } from 'igniteui-angular';
 // exported below as re-cast to create declaration type with expanded properties
 const ComboResourceStringsBG_: ExpandRequire<IComboResourceStrings> = {
     igx_combo_empty_message: 'Списъкът е празен',
-    igx_combo_enabledFilter_placeholder: 'Въведете термин за търсене',
+    igx_combo_filter_search_placeholder: 'Въведете термин за търсене',
     igx_combo_addCustomValues_placeholder: 'Добавяне на елемент',
     igx_combo_clearItems_placeholder: 'Изчистване на избора'
 };