+++ projects/igniteui-angular-i18n/src/i18n/DE/combo-resources.ts
@@ -3,7 +3,7 @@ import { IComboResourceStrings } from 'igniteui-angular';
 // exported below as re-cast to create declaration type with expanded properties
 const ComboResourceStringsDE_: ExpandRequire<IComboResourceStrings> = {
     igx_combo_empty_message: 'Die Liste ist leer',
-    igx_combo_enabledFilter_placeholder: 'Suchbegriff eingeben',
+    igx_combo_filter_search_placeholder: 'Suchbegriff eingeben',
     igx_combo_addCustomValues_placeholder: 'Element hinzufügen',
     igx_combo_clearItems_placeholder: 'Auswahl löschen'
 };