+++ projects/igniteui-angular-i18n/src/i18n/FR/combo-resources.ts
@@ -3,7 +3,7 @@ import { IComboResourceStrings } from 'igniteui-angular';
 // exported below as re-cast to create declaration type with expanded properties
 const ComboResourceStringsFR_: ExpandRequire<IComboResourceStrings> = {
     igx_combo_empty_message: 'La liste est vide',
-    igx_combo_enabledFilter_placeholder: 'Entrez un terme de recherche',
+    igx_combo_filter_search_placeholder: 'Entrez un terme de recherche',
     igx_combo_addCustomValues_placeholder: 'Ajouter un élément',
     igx_combo_clearItems_placeholder: 'Effacer la sélection'
 };