+++ projects/igniteui-angular-i18n/src/i18n/HU/combo-resources.ts
@@ -3,7 +3,7 @@ import { IComboResourceStrings } from 'igniteui-angular';
 // exported below as re-cast to create declaration type with expanded properties
 const ComboResourceStringsHU_: ExpandRequire<IComboResourceStrings> = {
     igx_combo_empty_message: 'Üres a lista',
-    igx_combo_enabledFilter_placeholder: 'Írjon be egy keresési kifejezést',
+    igx_combo_filter_search_placeholder: 'Írjon be egy keresési kifejezést',
     igx_combo_addCustomValues_placeholder: 'Elem hozzáadása',
     igx_combo_clearItems_placeholder: 'Kiválasztás törlése'
 };