+++ projects/igniteui-angular-i18n/src/i18n/JA/combo-resources.ts
@@ -3,7 +3,7 @@ import { IComboResourceStrings } from 'igniteui-angular';
 // exported below as re-cast to create declaration type with expanded properties
 const ComboResourceStringsJA_: ExpandRequire<IComboResourceStrings> = {
     igx_combo_empty_message: 'リストが空です',
-    igx_combo_enabledFilter_placeholder: '検索条件の入力',
+    igx_combo_filter_search_placeholder: '検索条件の入力',
     igx_combo_addCustomValues_placeholder: '項目の追加',
     igx_combo_clearItems_placeholder: '選択のクリア'
 };