+++ projects/igniteui-angular-i18n/src/i18n/PT/combo-resources.ts
@@ -3,7 +3,7 @@ import { IComboResourceStrings } from 'igniteui-angular';
 // exported below as re-cast to create declaration type with expanded properties
 const ComboResourceStringsPT_: ExpandRequire<IComboResourceStrings> = {
     igx_combo_empty_message: 'A lista está vazia',
-    igx_combo_enabledFilter_placeholder: 'Digite um termo de pesquisa',
+    igx_combo_filter_search_placeholder: 'Digite um termo de pesquisa',
     igx_combo_addCustomValues_placeholder: 'Adicionar item',
     igx_combo_clearItems_placeholder: 'Limpar seleção'
 };