+++ projects/igniteui-angular-i18n/src/i18n/ZH-HANS/combo-resources.ts
@@ -3,7 +3,7 @@ import { IComboResourceStrings } from 'igniteui-angular';
 // exported below as re-cast to create declaration type with expanded properties
 const ComboResourceStringsZHHANS_: ExpandRequire<IComboResourceStrings> = {
     igx_combo_empty_message: '列表为空',
-    igx_combo_enabledFilter_placeholder: '输入搜索字符串',
+    igx_combo_filter_search_placeholder: '输入搜索字符串',
     igx_combo_addCustomValues_placeholder: '添加项目',
     igx_combo_clearItems_placeholder: '清除选择'
 };